Backend: Rename wardrobe and equipment API classes for clarity
