+++ src/main/java/at/hannibal2/skyhanni/compat/SkyHanniReiPlugin.kt
@@ -2,7 +2,7 @@ package at.hannibal2.skyhanni.compat
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.features.inventory.wardrobe.CustomWardrobe
-import at.hannibal2.skyhanni.features.inventory.wardrobe.ArmorWardrobeApi
+import at.hannibal2.skyhanni.features.inventory.wardrobe.WardrobeApi
 import me.shedaniel.math.Rectangle
 import me.shedaniel.rei.api.client.plugins.REIClientPlugin
 import me.shedaniel.rei.api.client.registry.screen.ExclusionZones
@@ -12,7 +12,7 @@ class SkyHanniReiPlugin : REIClientPlugin {
 
     override fun registerExclusionZones(zones: ExclusionZones) {
         zones.register(ContainerScreen::class.java) { screen ->
-            if (ArmorWardrobeApi.inCustomWardrobe) {
+            if (WardrobeApi.inCustomWardrobe) {
                 if (CustomWardrobe.renderableTopCorner == Pair(0, 0)) {
                     listOf(screen.fullRectangle())
                 } else {
