+++ src/main/java/at/hannibal2/skyhanni/features/inventory/loadout/LoadoutApi.kt
@@ -5,7 +5,7 @@ import at.hannibal2.skyhanni.api.event.HandleEvent
 import at.hannibal2.skyhanni.data.ProfileStorageData
 import at.hannibal2.skyhanni.events.InventoryOpenEvent
 import at.hannibal2.skyhanni.events.InventoryUpdatedEvent
-import at.hannibal2.skyhanni.features.inventory.EquipmentApi
+import at.hannibal2.skyhanni.features.inventory.CurrentEquipmentApi
 import at.hannibal2.skyhanni.features.inventory.EquipmentSlot
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.utils.DelayedRun
@@ -195,7 +195,7 @@ object LoadoutApi {
         data.hotm = itemsList[HOTM_SLOT].parseCurrentSelection()
         data.hotf = itemsList[HOTF_SLOT].parseCurrentSelection()
 
-        EquipmentSlot.entries.forEach { EquipmentApi.setEquipment(it, data.equipment[it.ordinal]) }
+        EquipmentSlot.entries.forEach { CurrentEquipmentApi.setEquipment(it, data.equipment[it.ordinal]) }
     }
 
     // This is for Hotm, Hotf and Powerstone
